# docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/admin/receipts:
    get:
      summary: Query indexed receipts
      description: Query receipts indexed off-chain (no Soroban call).
      operationId: queryReceipts
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: dealId
          in: query
          required: false
          schema:
            type: string
        - name: txType
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paged receipts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedReceipts'
        default:
          $ref: "#/components/responses/Error"
  /api/deals/{dealId}/receipts:
    get:
      summary: List receipts for a deal
      description: Returns all receipts for a deal.
      operationId: listDealReceipts
      tags:
        - Deals
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Receipt list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealReceiptsResponse'
        default:
          $ref: "#/components/responses/Error"
...
components:
  schemas:
...
    IndexedReceipt:
      type: object
      required:
        - txId
        - txType
        - dealId
        - amountUsdc
        - externalRefHash
        - ledger
        - indexedAt
      properties:
        txId:
          type: string
        txType:
          type: string
        dealId:
          type: string
        listingId:
          type: string
        amountUsdc:
          type: string
        amountNgn:
          type: number
        fxRate:
          type: number
        fxProvider:
          type: string
        from:
          type: string
        to:
          type: string
        externalRefHash:
          type: string
        metadataHash:
          type: string
        ledger:
          type: integer
        indexedAt:
          type: string
          format: date-time
    PagedReceipts:
      type: object
      required:
        - data
        - total
        - page
        - pageSize
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/IndexedReceipt'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    DealReceiptsResponse:
      type: object
      required:
        - dealId
        - receipts
        - total
      properties:
        dealId:
          type: string
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/IndexedReceipt'
        total:
          type: integer
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...